Added verification on the delete asset method if the user trying to delete the asset is the owner of the asset. If not, the user cannot delete the asset.

+++ portfolio-manager-core/src/main/java/com/portfolio/portfoliomanager/usecases/DeleteAssetUseCase.java
@@ -1,6 +1,8 @@
 package com.portfolio.portfoliomanager.usecases;
 
 import com.portfolio.portfoliomanager.dataprovider.PortfolioManagerDataProvider;
+import com.portfolio.portfoliomanager.domain.Asset;
+import com.portfolio.portfoliomanager.exception.AuthorizationFailedException;
 import lombok.*;
 import org.springframework.stereotype.Service;
 
@@ -9,8 +11,26 @@ import org.springframework.stereotype.Service;
 public class DeleteAssetUseCase {
 
     private final PortfolioManagerDataProvider dataProvider;
+    private final FindAssetByExternalIdUseCase findAssetByExternalIdUseCase;
 
     public void execute(Input input) {
+        Asset asset = findAssetByExternalIdUseCase.execute(
+                FindAssetByExternalIdUseCase.Input.builder()
+                        .digitalUserId(input.getDigitalUserId())
+                        .externalId(input.getExternalId())
+                        .build()
+        ).getAsset();
+
+        if (!asset.getPermissionPolicy().equals(Asset.PermissionPolicy.OWNER)) {
+            throw new AuthorizationFailedException(
+                    String.format(
+                            "Digital user %s cannot delete expense %s because it is not its owner",
+                            input.getDigitalUserId(),
+                            input.getExternalId()
+                    )
+            );
+        }
+
         dataProvider.deleteAsset(input.getDigitalUserId(), input.getExternalId());
     }
